+++ next.config.js
@@ -20,26 +20,23 @@ module.exports = (phase, { defaultConfig }) => {
   };
 };
 
-// 외부에서 이미지를 불러올때 next.config.js에서 외부 경로를 설정해주어야한다. 
+// 외부에서 이미지를 불러올때 next.config.js에서 외부 경로를 설정해주어야한다.
 module.exports = {
   images: {
-    domains: [ 'www.kopis.or.kr' ],
+    domains: ['www.kopis.or.kr'],
     unoptimized: true,
-    imageSizes: [300, 640, 750, 828, 1000, 1200, 1920, 2048, 2840]
-  }
-}
-
-module.exports = {
-    async rewrites() {
-        return [
-            // {
-            //     destination: `http://localhost:4000`,
-            //     source: "/main/:image*",
-            // },
-            {
-                destination: `http://www.kopis.or.kr/openApi/restful/pblprfr`,
-                source: "/get",
-            },
-        ];
-    },
+    imageSizes: [300, 640, 750, 828, 1000, 1200, 1920, 2048, 2840],
+  },
+  async rewrites() {
+    return [
+      {
+        destination: `http://localhost:4000`,
+        source: '/main/:image*',
+      },
+      {
+        destination: `http://www.kopis.or.kr/:path*`,
+        source: '/openApi/restful/:path*',
+      },
+    ];
+  },
 };
