+++ pages/mainPages/Performance.jsx
@@ -1,9 +1,9 @@
 /** @jsxImportSource @emotion/react **/
 import { css } from '@emotion/react';
-import PerformanceList from '@/src/components/molecules/PerformanceList';
+import PerformanceList from '../../src/components/molecules/PerformanceList';
 import { useState, useEffect } from 'react';
-import Loading from '@/src/components/atoms/Loading';
-import useGetPerformance from '@/src/store/server/useGetPerformance';
+import Loading from '../../src/components/atoms/Loading';
+import useGetPerformance from '../../src/store/server/useGetPerformance';
 import { DateRange } from 'react-date-range';
 import 'react-date-range/dist/styles.css'; // main style file
 import 'react-date-range/dist/theme/default.css'; // theme css file
