+++ pages/subPages/Bookmark.jsx
@@ -1,26 +1,22 @@
-// localStorage로 저장하기
-
-import PerformanceList from '@/src/components/molecules/PerformanceList';
-import { getBookmarkInfo } from '@/src/modules/bookmarkModules';
-import { async } from '@firebase/util';
+import PerformanceList from '../../src/components/molecules/PerformanceList';
+import { getBookmarkInfo } from '../../src/modules/bookmarkModules';
 import { useQuery } from 'react-query';
-import { useEffect, useState } from 'react';
 import axios from 'axios';
 
+// localStorage로 저장하기
 export const getListApi = async () => {
-    const list = getBookmarkInfo();
-    const urlArr = [];
-    list.forEach((elem) => {
-        const url = `http://localhost:4000/description/${elem}`;
-        urlArr.push(axios.get(url));
-    });
-    const response = await axios.all(urlArr);
+  const list = getBookmarkInfo();
+  const urlArr = [];
+  list.forEach((elem) => {
+    const url = `http://localhost:4000/description/${elem}`;
+    urlArr.push(axios.get(url));
+  });
+  const response = await axios.all(urlArr);
 
   // 파싱
   const arr = [];
   response.map((elem) => {
     const elemData = elem.data.elements[0].elements[0].elements;
-    // console.log(elemData);
     const obj = {
       id: elemData[0].elements[0]?.text || '정보없음',
       name: elemData[1].elements[0]?.text || '정보없음',
@@ -38,8 +34,6 @@ export const getListApi = async () => {
 
 const Bookmark = () => {
   const { data, isLoading } = useQuery(['listApi'], getListApi);
-
-  // console.log(data);
 
   if (isLoading) {
     return <div>로딩중</div>;
