--- src/components/Recoil/recoil-addReservation.js
@@ -1,12 +1,0 @@
-import useGetDescription from '@/src/store/server/useGetDescription';
-import { atom, selector } from 'recoil';
-
-export const performanceListState = atom({
-  key: 'performanceListState',
-  default: [],
-});
-
-export const performanceState = atom({
-  key: 'performanceState',
-  default: '',
-});
