+++ src/components/molecules/reservation/Reservation.jsx
@@ -1,0 +1,11 @@
+import ReservationItem from './ReservationItem';
+
+const Reservation = ({ data }) => {
+  return (
+    <div>
+      <ReservationItem total={data} />
+    </div>
+  );
+};
+
+export default Reservation;
